# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API de manipulação de produtos",
    "description": "Essa API tem como objetivo a manipulação dos produtos",
    "termsOfService": "http://localhost:3000/terms",
    "contact": {
      "email": "rafael_santos.s1@hotmail.com"
    },
    "version": "1.0.0"
  },
  "servers":[
    {
      "url": "http://localhost:3000/v1",
      "description": "API de test"
    }
  ],
  "paths": {
    "/products": {
      "post": {
        "summary": "Cadastro de produto",
        "description": "Essa rota será responsável por cadastrar um novo produto",
        "tags": ["Products"],
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Product"
              },
              "examples": {
                "product": {
                  "value": {
                    "name": "Teclado RGB",
                    "description": "Teclado Gamer com RGB",
                    "price": 680.90
                  }
                }
              }
            }
          }
        },
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "400":{
            "description": "Product Already exists!"
          }, 
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          }
        }
      }
    },
    "/": {
      "get": {
        "description": "Retorna todos os produtos cadastrados na BD",
        "summary": "Retorna todos os produtos cadastrados na BD",
        "tags": ["Products"],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "itens": {
                    "$ref": "#/components/schemas/Product"
                  }
                },
                "examples": {
                  "product": {
                    "value": {
                      "name": "Teclado RGB",
                      "description": "Teclado Gamer com RGB",
                      "price": 680.90
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/products/findByName": {
      "get": {
        "description": "Busca de um produto pelo nome",
        "summary": "Busca de um produto pelo nome",
        "tags": ["Products"],
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "description": "Nome do produto para busca",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "itens": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/products/{id}": {
      "get": {
        "description": "Busca de um produto pelo id",
        "summary": "Busca de um produto pelo id",
        "tags": ["Products"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Nome do produto pelo ID",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          }
        }
      },
      "put": {
        "description": "Alterar um produto pelo id",
        "summary": "Alterar um produto pelo id",
        "tags": ["Products"],
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Nome do produto pelo ID",
            "required": true
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Product"
              },
              "examples": {
                "product": {
                  "value": {
                    "name": "Teclado RGB sem fio",
                    "description": "Teclado Gamer com RGB sem fio",
                    "price": 1280.90
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Product": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "id": {
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}